% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contiguous_congressional_districts_spdf}
\alias{contiguous_congressional_districts_spdf}
\title{Contiguous US Congressional Districts 2014}
\format{A large SpatialPolygonsDataFrame (433 elements, 10.9Mb), an internal data.frame called "data" is available with the following information:
\describe{
  \item{state.fp}{STATEFP, State FIP code}
  \item{cd114fp}{Unknown identifier from Census}
  \item{affgeoid}{American Fact Finder ID}
  \item{geoid}{Unknown identifier from Census}
  \item{lsad}{Unknown identifier from Census}
  \item{cdsessn}{Unknown identifier from Census}
  \item{a.land}{Area of state that is land}
  \item{a.water}{Area of state that is water}
  \item{state.name}{Name of state district belongs to}
  \item{namelsad}{Name of congressional district}
  \item{name}{Canonical name of district}
}}
\source{
\url{http://www2.census.gov/geo/tiger/GENZ2014/shp/cb_2014_us_cd114_500k.zip}
}
\usage{
contiguous_congressional_districts_spdf
}
\description{
A SpatialPolygonsDataFrame representing the congressional districts of the contiguous United States of America with 500k = 1:500,000 resolution, from www.census.gov.
}
\keyword{datasets}

