% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contiguous_us_shapefiles}
\alias{contiguous_us_shapefiles}
\title{Contiguous US Shapefiles}
\format{A large SpatialPolygonsDataFrame (49 elements, 3.7Mb), an internal data.frame called "data" is available with the following information:
\describe{
  \item{state.fp}{STATEFP, State FIP code}
  \item{state.ns}{GNIS ID for State}
  \item{state.short.name}{Two-character short name for state}
  \item{state.name}{State name}
  \item{a.land}{Area of state that is land}
  \item{a.water}{Area of state that is water}
}}
\source{
\url{https://www.census.gov/geo/maps-data/data/cbf/cbf_state.html}
}
\usage{
contiguous_us_shapefiles
}
\description{
A SpatialPolygonsDataFrame representing the contiguous United States of America with 500k = 1:500,000 resolution, from www.census.gov.
}
\keyword{datasets}

