% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips_codes}
\alias{fips_codes}
\title{FIPS Codes}
\format{A DataFrame (57 rows, 5 columns) with the following data:
\describe{
  \item{fips.code}{State FIP code}
  \item{state.short.name}{Two-character short name for state}
  \item{state.name}{State name}
  \item{state.ns}{GNIS ID for State}
  \item{contiguous.united.states}{Member of contiguous United States or note (TRUE or FALSE)}
}}
\source{
\url{https://www.census.gov/geo/maps-data/data/cbf/cbf_state.html}
}
\usage{
fips_codes
}
\description{
A DataFrame containing the FIPS codes for states, taken from https://www.census.gov/geo/reference/codes/cou.html. Whether a State is included in the contiguous United States or not was entered by hand, raise issues in the issue tracker if there are any disagreements.
}
\keyword{datasets}

