% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-shapefiles.R
\name{get_us_shape_files}
\alias{get_us_shape_files}
\title{get_shape_files}
\usage{
get_us_shape_files(resolution = "500k", year = "2015",
  zip.path = "shapefiles.zip", shapefiles.path = "shapefiles/",
  delete.zip = TRUE)
}
\arguments{
\item{resolution}{Resolution of the requested shapefiles, may only be "500k", "5m", or "20m". Default to "500k"}

\item{year}{Year for which the shapefiles should be requested, may only be 2013, 2014, or 2015. Default to 2015}

\item{zip.path}{Path for zip file downloaded from census.gov, default to shapefiles.zip in the working directory.}

\item{shapefiles.path}{Path for shapefiles to be unzipped, default to shapefiles/}

\item{delete.zip}{Delete zip or not, default to TRUE.}
}
\description{
\code{get_us_shape_files} downloads shapefiles from census.gov
}

