% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-shapefiles.R
\name{make_us_contiguous}
\alias{make_us_contiguous}
\title{make_us_contiguous}
\usage{
make_us_contiguous(us_spdf)
}
\arguments{
\item{us_spdf}{SpatialPolygonDataFeame for the United States, requires @data with column state.fips}
}
\description{
\code{make_us_contiguous} removes states that are not in the contiguous United States, and drops levels from the internal data object.
}

