% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-shapefiles.R
\name{make_us_spdf}
\alias{make_us_spdf}
\title{make_us_spdf}
\usage{
make_us_spdf(shapefiles.path = "shapefiles/", year = "2015",
  resolution = "500k", contiguous.only = TRUE, download.shapefiles = TRUE)
}
\arguments{
\item{shapefiles.path}{Path for shapefiles, if \code{download.shapefiles=FALSE} these must already be downloaded. Default to "shapefiles/"}

\item{year}{Year for which the shapefiles should be requested, may only be 2013, 2014, or 2015. If \code{download.shapefiles=FALSE} these must already be downloaded. Default to 2015}

\item{resolution}{Resolution of the shapefiles, may only be "500k", "5m", or "20m". If \code{download.shapefiles=FALSE} these must already be downloaded. Default to "500k"}

\item{contiguous.only}{If TRUE the returned SPDF will only contain States in the contiguous United States. Default to TRUE.}

\item{download.shapefiles}{Download shapefiles in this function call, rather than separately calling \code{get_us_shape_files}. Default to TRUE}
}
\description{
\code{make_us_spdf} returns shape files from census.gov
}

