% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyr_less_mapvalues.R
\name{plyr_less_mapvalues}
\alias{plyr_less_mapvalues}
\title{plyr_less_mapvalues}
\usage{
plyr_less_mapvalues(x, from, to, warn_missing = TRUE)
}
\description{
\code{plyr_less_mapvalues} provides mapvalues from plyr without conflicts arising from the rest of the tidyverse
}
\note{
see plyr documentation
}

