% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapefile_details}
\alias{shapefile_details}
\title{Details of the shapefiles included in this package.}
\format{A tibble containing 6 rows and 3 variables.
\describe{
    \item{description}{What is this shapefile used for?}
    \item{year}{Year of observation (i.e. congressional districts for 2016)}
    \item{url}{URL for the shapefile zip file}
}}
\source{
\url{http://www2.census.gov/geo/tiger/GENZ2016/shp/cb_2016_us_state_20m.zip}
}
\usage{
shapefile_details
}
\description{
A tibble containing the url, year and description of all shapefiles used to create the datasets included in this package
}
\keyword{datasets}
