% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statesrcontiguous_shapefile_details}
\alias{statesrcontiguous_shapefile_details}
\title{Details of the shapefiles included in this package.}
\format{A tibble containing 6 rows and 3 variables.
\describe{
    \item{path}{Internal package path for file containing url and update year}
    \item{subdivision}{Which subdivision does this data pertain to; States, Congressional Districts or Counties}
    \item{year}{Year of observation (i.e. congressional districts for 2016)}
    \item{source}{URL for the data}
}}
\source{
\url{http://www2.census.gov}
}
\usage{
statesrcontiguous_shapefile_details
}
\description{
A tibble containing the url, year and description of all shapefiles and additional datasets used to create the objects included in this package
}
\keyword{datasets}
